JdkMarshaller turned to interface
