+++ modules/binary/api/src/main/java/org/apache/ignite/marshaller/jdk/JdkMarshallerDummySerializable.java
