+++ modules/core/src/main/java/org/apache/ignite/internal/marshaller/optimized/OptimizedClassDescriptor.java
@@ -46,8 +46,6 @@ import java.util.Properties;
 import java.util.TreeMap;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentMap;
-
-import org.apache.ignite.internal.binary.GridBinaryMarshaller;
 import org.apache.ignite.internal.util.GridUnsafe;
 import org.apache.ignite.internal.util.SerializableTransient;
 import org.apache.ignite.internal.util.typedef.F;
@@ -55,6 +53,7 @@ import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.lang.IgniteProductVersion;
 import org.apache.ignite.marshaller.MarshallerContext;
 import org.apache.ignite.marshaller.MarshallerExclusions;
+import org.apache.ignite.marshaller.Marshallers;
 
 import static java.lang.reflect.Modifier.isFinal;
 import static java.lang.reflect.Modifier.isPrivate;
@@ -754,7 +753,7 @@ class OptimizedClassDescriptor {
             case OBJ_ARR:
                 OptimizedClassDescriptor compDesc = OptimizedMarshallerUtils.classDescriptor(clsMap,
                     obj.getClass().getComponentType(),
-                    GridBinaryMarshaller.USE_CACHE.get(),
+                    Marshallers.USE_CACHE.get(),
                     ctx,
                     mapper);
 
@@ -816,7 +815,7 @@ class OptimizedClassDescriptor {
 
             case CLS:
                 OptimizedClassDescriptor clsDesc = OptimizedMarshallerUtils.classDescriptor(
-                    clsMap, (Class<?>)obj, GridBinaryMarshaller.USE_CACHE.get(), ctx, mapper);
+                    clsMap, (Class<?>)obj, Marshallers.USE_CACHE.get(), ctx, mapper);
 
                 clsDesc.writeTypeData(out);
 
@@ -827,7 +826,7 @@ class OptimizedClassDescriptor {
 
                 for (Class<?> intf : proxyIntfs) {
                     OptimizedClassDescriptor intfDesc = OptimizedMarshallerUtils.classDescriptor(
-                        clsMap, intf, GridBinaryMarshaller.USE_CACHE.get(), ctx, mapper);
+                        clsMap, intf, Marshallers.USE_CACHE.get(), ctx, mapper);
 
                     intfDesc.writeTypeData(out);
                 }
